move service loading to docker domain
